feat - finish unit testing

+++ Backend/src/main/java/com/firomsa/ecommerce/repository/ImageRepository.java
@@ -8,9 +8,10 @@ import org.springframework.stereotype.Repository;
 import com.firomsa.ecommerce.model.Image;
 
 @Repository
-public interface ImageRepository extends JpaRepository<Image, Long>{
+public interface ImageRepository extends JpaRepository<Image, Integer> {
 
     Optional<Image> findByName(String id);
+
     boolean existsByName(String id);
-    
+
 }
